Added: Refactoring the code and check validation and connect with microservices

+++ fundoo-notes/src/main/java/com/project/fundoonotes/controller/ResponseController.java
@@ -28,7 +28,7 @@ public class ResponseController {
      * @return : id
      */
 	@GetMapping("/getAllData/{id}")
-	public ResponseTemplateDto getLablesAndNotesByUser(@PathVariable("id") long userId) {
+	public ResponseTemplateDto getLablesAndNotesByUser(@PathVariable("id") int userId) {
 		
 		return ResponseService.getLablesAndNotesWithUser(userId);
 	}
