+++ fundoo-notes/src/main/java/com/project/fundoonotes/repo/LabelRepository.java
@@ -16,6 +16,6 @@ public interface LabelRepository extends JpaRepository<Label, Integer> {
 
 	Label save(LabelDto labelDto);
 
-	List<Label> findByUserId(long userId);
+	List<Label> findByUserId(int userId);
 
 }
